Factor device structure out into common package
Thus improves dependency graph and helps avoid circular dependencies
